backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=13.0
signalrcore>=0.9.5
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.10.0
python-dateutil>=2.8.2
